requirements.txt
discord.py
yfinance
requests
python-dotenv
apscheduler